backend/requirements.txt
Django
psycopg2
django-rest-framework
django-reset-migrations
pytest
pathlib
pytest-django
django-safedelete